refactor(sonnet): 기존 시스템 Container 패턴 적용
Container Provider 의존성 주입 통합
- App 및 Container 초기화 로직 업데이트
- Handler 팩토리 및 레지스트리 Container 적용
- 메트릭 시스템 Container 통합
- 미들웨어 및 라우터 Container 지원

+++ metrics/ttl_collector_test.go
@@ -121,7 +121,7 @@ func TestTTLCollector_ExpirationTracking(t *testing.T) {
 }
 
 func TestCalculatePercentile(t *testing.T) {
-	values := []int{100, 200, 300, 400, 500, 600, 700, 800, 900, 1000}
+	values := []float64{100, 200, 300, 400, 500, 600, 700, 800, 900, 1000}
 
 	// 50th percentile (median) should be 550
 	p50 := calculatePercentile(values, 50)
